Factor tests into a new file
